Going with a somewhat simplistic error strategy
The parser doesn't let the parsing code directly manipulate it -
instead we only expose functions to move the parse state forward
given certain conditions.  E.g. consume a token if it's kind X.

But we also expose an 'expect' function which when we do not have the
expected token we produce an error & consume the token.  We also set the
parser to be in a panicking state.

If whilst panicking we encounter further unexpected tokens we do not
produce further error syntax nodes, just tombstone tokens.

If we encounter an expected token we presume we're back on track and
exit panicking.