I need to take a closer look at begin/complete binding status
I'm not really using it, and it seems that I'm still detecting
the edges cases that that data supposedly helps us catch.

e.g. in Lox it was using a closed value to initialize a
variable of the same name.  Not technically wrong, just
potentially confusing.

let x = 42;
{
 let x = x + 1;
}

Lox also bans shadowing, e.g.
let x = 42;
let x = 32; // wrong

We've accomplished both of these, and only seem to need a Set.