We need to solve 2 issues for functions:
1. How to obtain a reference to a subset of the syntax tree for the newt-callable
2. How to return from a call
3. Where to store side information

Answer #1: We're going to make the tree cloneable.  Later we can probably
transition to Rc.

Answer #2: We're going to re-jigger the control flow of the visitor to be
similar to how the visitor in gen-syntax works.  Then we'll add a return 'error'
which is actually handled.

Answer #3: We're going to continue with side tables for now.  But I have a notion
of moving SyntaxNodes over to Vec & we can add children at runtime.