More progress, I need to reconsider how I mutate my tree.
Perhaps if I have a Green tree (IIRC this is the long lived one in other compilers)
which is mutable & singly owned (boxed vec), and a Red tree which is immutable but
can share ownership (rc slice).  Cloning from a Green to a Red is straightforward enough,
but how to make both navigable by visitors?  That's a lot of duplicated code.